% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\name{Atkinson}
\alias{Atkinson}
\title{Atkinson}
\usage{
Atkinson(X, W = rep(1, length(X)), e = 1)
}
\arguments{
\item{X}{is a data vector}

\item{W}{is a vector of weights}

\item{e}{is a parameter for calculating the value of the Atkinson coefficient}
}
\value{
The value of Atkinson coefficient.
}
\description{
Computes Atkinson inequality measure of a given variable taking into account weights.
}
\examples{
X=c(1,2,3,4,5,6,7,8,9)
W=c(2,5,6,7,3,4,5,2,5)
Atkinson(X,W)

}
\references{
Philip B. Coulter: (1989) Measuring Inequality ISBN 0-8133-7726-9
}
