% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\name{CEO_to_Worker}
\alias{CEO_to_Worker}
\title{CEO-to-Worker Pay Ratio (in general: max-to-median)}
\usage{
CEO_to_Worker(X, W = rep(1, length(X)))
}
\arguments{
\item{X}{is a data vector}

\item{W}{is a vector of weights}
}
\value{
The value of maximum-to-median ratio, in particular: CEO-to-Worker Pay Ratio for salary data.
}
\description{
CEO-to-Worker Pay Ratio - it is a pay ratio between CEO pay and median employee pay.
}
\details{
CEO-to-Worker Pay Ratio is generalized to any numeric data vector by the following formula:
\deqn{CEO_to_Worker =\frac{H}{L}}
where \eqn{H} is the maximum value and
\eqn{L} is the median value.
Since 2018, U.S. publicly held corporations have had to annually report
the ratio between their CEO and median worker compensation.
}
\examples{
# Simple example for fake salary data
X=seq.int(5000,15000,1000)
W=11:1
CEO_to_Worker(X,W)

data(Tourism)
# Max-to-median (CEO_to_Worker) for total expenditure with sample weights
X=Tourism$Total_expenditure
W=Tourism$Sample_weight
CEO_to_Worker(X,W)



}
