% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\name{CoefVar}
\alias{CoefVar}
\title{CoefVar}
\usage{
CoefVar(X, W = rep(1, length(X)), square = FALSE, na.rm = TRUE)
}
\arguments{
\item{X}{is a data vector}

\item{W}{is a vector of weights}

\item{square}{logical, argument of the function CoefVar, for details see below}

\item{na.rm}{logical, should missing values (NAs) be removed prior to computations? If set to FALSE the computations yield NA}
}
\value{
The value of CoefVar coefficient.
}
\description{
Computes CoefVar inequality measure of a given variable taking into account weights.
}
\examples{
X=c(1,2,3,4,5,6,7,8,9)
W=c(2,5,6,7,3,4,5,2,5)
CoefVar(X,W)

}
\references{
Philip B. Coulter: (1989) Measuring Inequality ISBN 0-8133-7726-9
}
