% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\name{Jenkins}
\alias{Jenkins}
\title{Jenkins and Cowell_and_Flachaire}
\usage{
Jenkins(X, W = rep(1, length(X)), alfa = 0.8)
}
\arguments{
\item{X}{is a data vector}

\item{W}{is a vector of weights}

\item{alfa}{is the Jenkins coefficient parameter}
}
\value{
The value of Jenkins and Cowell_and_Flachaire coefficient.
}
\description{
Computes Jenkins and Cowell_and_Flachaire inequality measure of a given variable taking into account weights.
}
\examples{
X=c(1,2,3,4,5,6,7,8,9)
W=c(2,5,6,7,3,4,5,2,5)
Jenkins(X,W)

}
\references{
Philip B. Coulter: (1989) Measuring Inequality ISBN 0-8133-7726-9
F A Cowell: Measurement of Inequality, 2000, in A B Atkinson / F Bourguignon (Eds): Handbook of Income Distribution, Amsterdam
}
