% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\name{Kolm}
\alias{Kolm}
\title{Kolm}
\usage{
Kolm(X, W = rep(1, length(X)), parameter = 1, na.rm = TRUE)
}
\arguments{
\item{X}{is a data vector}

\item{W}{is a vector of weights}

\item{parameter}{is a Kolm parameter}

\item{na.rm}{logical, should missing values (NAs) be removed prior to computations? If set to FALSE the computations yield NA}
}
\value{
The value of Kolm coefficient.
}
\description{
Computes Kolm inequality measure of a given variable taking into account weights.
}
\examples{
X=c(1,2,3,4,5,6,7,8,9)
W=c(2,5,6,7,3,4,5,2,5)
Kolm(X,W)

}
\references{
Philip B. Coulter: (1989) Measuring Inequality ISBN 0-8133-7726-9
}
