% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\name{Leti}
\alias{Leti}
\title{Leti}
\usage{
Leti(X, W = rep(1, length(X)))
}
\arguments{
\item{X}{is a data vector}

\item{W}{is a vector of weights}
}
\value{
The value of Leti coefficient.
}
\description{
Computes Leti inequality measure of a given variable taking into account weights.
}
\examples{
X=c(1,2,3,4,5,6,7,8,9)
W=c(2,5,6,7,3,4,5,2,5)
Leti(X,W)

}
\references{
Philip B. Coulter: (1989) Measuring Inequality ISBN 0-8133-7726-9
}
