% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\name{Prop20_20}
\alias{Prop20_20}
\title{Prop20:20}
\usage{
Prop20_20(X, W = rep(1, length(X)))
}
\arguments{
\item{X}{is a data vector}

\item{W}{is a vector of weights}
}
\value{
The value of 20:20 ratio coefficient.
}
\description{
20:20 ratio - the ratio of thr total income of the 20\% richest people to the 20\% poorest people.
}
\examples{
X=c(1,2,3,4,5,6,7,8,9)
W=c(2,5,6,7,3,4,5,2,5)
Prop20_20(X,W)

}
\references{
Philip B. Coulter: (1989) Measuring Inequality ISBN 0-8133-7726-9
Panel Data Econometrics: Theoretical Contributions And Empirical Applications edited by Badi Hani Baltag
Notes on Statistical Sources and Methods - The Equality Trust.
}
