% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\name{RicciSchutz}
\alias{RicciSchutz}
\title{RicciSchutz}
\usage{
RicciSchutz(X, W = rep(1, length(X)), na.rm = TRUE)
}
\arguments{
\item{X}{is a data vector}

\item{W}{is a vector of weights}

\item{na.rm}{logical, should missing values (NAs) be removed prior to computations? If set to FALSE the computations yield NA}
}
\value{
The value of RicciSchutz coefficient.
}
\description{
Computes RicciSchutz inequality measure of a given variable taking into account weights.
}
\examples{
X=c(1,2,3,4,5,6,7,8,9)
W=c(2,5,6,7,3,4,5,2,5)
RicciSchutz(X,W)

}
\references{
Philip B. Coulter: (1989) Measuring Inequality ISBN 0-8133-7726-9
}
