% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\docType{data}
\name{Tourism}
\alias{Tourism}
\title{Sample survey on trips}
\format{
A data frame with 5319 observations of 17 variables
\itemize{
\item Year
\item Country
\item Country code
\item World region
\item Purpose of trip
\item Accommodation type
\item Number of trip's participants
\item Nights spent
\item Travel agency (organiser)
\item Sample weight
\item Total expenditure
\item Expenditure for organiser
\item Private expenditure
\item Expenditure on accommodation
\item Expenditure on restaurants & café
\item Expenditure on transport
\item Expenditure on commodities
}
}
\usage{
data(Tourism)
}
\description{
Data from sample survey on trips conducted in Polish households.
}
\details{
Answers were modified due to disclosure control. Data presents only part of full database.
}
\keyword{datasets}
