% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\docType{data}
\name{Well_being}
\alias{Well_being}
\title{Sample survey on quality of life}
\format{
A data frame with 1197 observations of 27 variables
\itemize{
\item Area. Rural and urban
\item Gender. Male and female
\item Year. Year of survey (2015 and 2019)
\item V1. I have good opportunities to use my talents and skills at work
\item V2. I am treated with respect by others at work
\item V3. I have adequate opportunities for vacations or leisure activities
\item V4. The quality of local services where (I) live is good
\item V5. There is very little pollution from cars or other sources where I spend most of my time
\item V6. There are parks and green areas near my residence
\item V7. I have the freedom to plan my life the way I want to
\item V8. I feel safe walking around my neighborhood during the day
\item V9. Overall, to what extent are you currently satisfied with your life
\item V10. Overall, to what extent do you feel that the things you do in life are worthwhile
\item V11. How do you rate your health
\item V12. How do you rate your work
\item V13. How do you rate your sleep
\item V14. How do you rate your leisure time
\item V15. How do you rate your family life
\item V16. How do you rate your community and public affairs life
\item V17. How do you rate your personal plans
\item V18. How do you rate your housing conditions
\item V19. How do you rate your personal income
\item V20. How do you rate your personal prospects
\item V21. Does being part of the local community make you feel good about yourself
\item V22. Do you have a say in what the local community is like
\item V23. Is your neighborhood a good place for you to live
\item Weight. Sample weight for each household
}
}
\usage{
data(Well_being)
}
\description{
Data from sample survey on quality of life conducted on Polish-Ukrainian border in 2015 and 2019.
}
\details{
Questions are on Likert scale: 1 - the worst assessment, 5 - the best assessment.
Only 23 questions were selected out of over 100 questions.
Answers were modified due to disclosure control.
}
\keyword{datasets}
