% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\name{AF}
\alias{AF}
\title{Allison and Foster}
\usage{
AF(X, W = rep(1, length(X)))
}
\arguments{
\item{X}{is a data vector}

\item{W}{is a vector of weights}
}
\value{
The value of Allison and Foster coefficient.
}
\description{
Computes Allison and Foster inequality measure of a given variable taking into account weights.
}
\examples{
X=c(1,2,3,4,5,6,7,8,9)
W=c(2,5,6,7,3,4,5,2,5)
AF(X,W)

}
\references{
Philip B. Coulter: (1989) Measuring Inequality ISBN 0-8133-7726-9
}
