% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\name{Palma}
\alias{Palma}
\title{Palma}
\usage{
Palma(X, W = rep(1, length(X)))
}
\arguments{
\item{X}{is a data vector}

\item{W}{is a vector of weights}
}
\value{
The value of Palma coefficient.
}
\description{
Palma proportion - the ratio of the total income of the 10\% richest people to the 40\% poorest people.
}
\examples{
X=c(1,2,3,4,5,6,7,8,9)
W=c(2,5,6,7,3,4,5,2,5)
Palma(X,W)

}
\references{
Philip B. Coulter: (1989) Measuring Inequality ISBN 0-8133-7726-9
Putting the Gini Back in the Bottle? 'The Palma' as a Policy-Relevant Measure of Inequality
}
