% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequalityMeasure.R
\docType{data}
\name{Tourism}
\alias{Tourism}
\title{Sample survey on trips}
\format{
A data frame with 384 observations of 14 variables
\itemize{
\item Year
\item Group of countries
\item Purpose of trip
\item Accommodation type
\item Total expenditure
\item Expenditure for organiser
\item Private expenditure
\item Expenditure on accommodation
\item Expenditure on restaurants & café
\item Expenditure on transport
\item Expenditure on commodities
\item Number of trip's participants
\item Nights spent
\item Sample weight
}
}
\usage{
data(Tourism)
}
\description{
Data from sample survey on trips conducted in Polish households.
}
\details{
Answers were modified due to disclosure control. Data presents only part of full database.
}
\keyword{datasets}
